#ADD: admin:    Report in    add product Admin    Ban Reason + messenge to user from Admin Area user:    Del Product in Cart #Change:  OrderDetail model: OrderStatusID from int to int16(int to smallint)

+++ Areas/User/Models/OrderDetail.cs
@@ -10,7 +10,7 @@ namespace DACN_DVTRUCTUYEN.Areas.User.Models
         public string OrderID { get; set; }
         public string ProductID { get; set; }
         public string ProductOptionValue { get; set; }
-        public int OrderStatusID { get; set; }
+        public Int16 OrderStatusID { get; set; }
         public long Amount { get; set; }
     }
 }
